gol: Change extendArray signature to use 2D array pointer
It looks like using a pointer to a pointer with a dynamically allocated
array doesn't seem to work well with the alignment.

Since this only amounts to at max 800 bytes on the stack at a time, let's just
go with using an array initialized on the stack. This also vastly
simplifies book-keeping for the caller.

Update the test to use the new signature.